% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solast.R
\name{lastcall}
\alias{lastcall}
\title{Retrieve Last Call from History}
\usage{
lastcall(max.lookback = 100L)
}
\arguments{
\item{max.lookback}{integer(1L) how many lines of history to attempt to
parse back from the end of the history file}
}
\value{
language the last top level command, parsed
}
\description{
Returns the last call that was entered at the R prompt if it parsed
parsed successfully.  Because the history file may contain unparseable lines
and because it may not start from the beginning of an R session, it is not
possible to guarantee that we will be able to correctly retrieve the last
command.  This function will warn when it can retrieve a command but is not
certain the command is complete.
}
\author{
Brodie Gaslam
}
