\name{readClip}
\alias{readClip}
\title{Read clipboard regardless of OS}
\usage{
readClip()
}
\value{
character string containing text on the clipboard.
}
\description{
Different operating systems have different ways of handling
the clipboard. Given the frequency with which text is
copied to the clipboard to place in an answer on
StackOverflow, this utility is provided.
}

