% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soanswer.R
\name{soanswer}
\alias{soanswer}
\title{Turn an R expression into an SO answer}
\usage{
soanswer(expr, in_task_callback = FALSE)
}
\arguments{
\item{expr}{An R expression.}

\item{in_task_callback}{A logical value.  Not intended for direct use.}
}
\value{
Markdown output is printed to the console, written to the clipboard 
and silently returned as a \code{noquote} character vector.
}
\description{
Parrot the input expression and append the output with knitr-style comments,
all as markdown.  \code{soanswer} changes the output for a single expression.
}
\examples{
# Output for various types, explicitly calling soanswer
soanswer(sin(2 * pi))
soanswer(sleep)
soanswer(message("A message!"))
soanswer(warning("A warning!"))
soanswer(stop("An error!"))
}
\seealso{
\code{\link[base]{addTaskCallback}}
}
\author{
Richard Cotton
}
