% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solast.R
\name{solast}
\alias{solast}
\title{Capture Last Top Level Call And Output}
\usage{
solast(silent = FALSE, clip = TRUE, drop.curly.brace = TRUE, ...)
}
\arguments{
\item{silent}{logical(1L) set to TRUE to suppress screen display}

\item{clip}{logical(1L) set to TRUE (default) to attempt to write to system
clipboard}

\item{drop.curly.brace}{if the last call is a call to \code{\{}, drop it and
show only the contents (see examples)}

\item{...}{arguments to pass on to \code{\link{lastcall}}}
}
\value{
character, invisibly, the call and value together formatted for
  posting on SO
}
\description{
Captures the output from \code{.Last.value} along with the last expression
from the history file and formats them in a manner suitable for posting on
Stack Overflow.
}
\details{
If you the expression you want captured spans multiple lines, enclose them
in a call to \code{\{}.  The curly braces will be dropped from the display
to make it appear as if you typed all the commands sequentially into the R
prompt, though any interim output will not be captured.

Please note that the only output that is "captured" by this function is the
result of evaluating \code{.Last.value}.  If there were warning, errors, or
other screen output during the actual expression evaluation those will not
show up here.
}
\examples{
\dontrun{
1 + 1
solast()
{
  x <- runif(20)
  x + 3
}
solast()
}
}
\author{
Andanda Mahto, Brodie Gaslam
}
\seealso{
\code{\link{lastcall}} for limitations in capturing the last call
}

