\name{sorandf}
\alias{sorandf}
\title{Create random data.frame}
\usage{
  sorandf(rows = 10L, cols = c("race", "gender", "age"),
    names = make.names(cols))
}
\arguments{
  \item{rows}{integer, number of rows in data.frame}

  \item{cols}{types of columns}

  \item{names}{names of columns}
}
\description{
  Create random data.frame
}

