% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sotrunc.R
\name{sotrunc}
\alias{sotrunc}
\alias{sotrunc.data.frame}
\alias{sotrunc.default}
\alias{sotrunc.list}
\alias{sotrunc.matrix}
\title{Truncate the Output of Dataframes, Matrices and Lists}
\usage{
sotrunc(x, ncol = 10, nrow = 15, ...)

\method{sotrunc}{matrix}(x, ncol = 10, nrow = 15, ...)

\method{sotrunc}{data.frame}(x, ncol = 10, nrow = 15, ...)

\method{sotrunc}{list}(x, ncol = 10, nrow = 15, ...)

\method{sotrunc}{default}(x, ncol = 10, nrow = 15, ...)
}
\arguments{
\item{x}{An R object (usually a \code{data.frame}, \code{matrix}, or a 
\code{list} that contains >= 1 \code{data.frame}/\code{matrix}.}

\item{ncol}{The number of columns (including 3 columns displayed as ellipsis) 
to display.}

\item{nrow}{The number of rows (including 3 rows displayed as ellipsis) to 
display.}

\item{\ldots}{Other arguments passed to \code{sotrunc}.}
}
\value{
Prints a truncated display of the input.
}
\description{
A dispaly printing function that trunates \code{data.frame}/\code{matrix} by 
adding ellipsis \ldots to take the place of non-displayed columns and rows.

\code{sotrunc.matrix} - matrix method for \code{sotrunc}

\code{sotrunc.data.frame} - data.frame method for \code{sotrunc}

\code{sotrunc.list} - list method for \code{sotrunc}

\code{sotrunc.default} - Default method for \code{sotrunc}
}
\note{
This function is intended for truncated display.  Please do not use to 
gernarate your \href{http://sscce.org/}{SSCCE}
(Short, Self Contained, Correct (Compilable), Example).
}
\examples{
sotrunc(mtcars)
sotrunc(mtcars, nrow = 6)
sotrunc(mtcars, ncol = 6)
sotrunc(CO2)
sotrunc(lm(mpg~hp+am, data=mtcars))
sotrunc(1:10)
sotrunc(mtcars[1:5, 1:5])
}
\keyword{truncate}

