% Generated by roxygen2 (4.0.0): do not edit by hand
\name{Truncate}
\alias{Truncate}
\alias{Truncate.data.frame}
\alias{Truncate.default}
\alias{Truncate.list}
\alias{Truncate.matrix}
\title{Truncate the Output of Dataframes, Matrices and Lists}
\usage{
Truncate(x, ncol = 10, nrow = 15, ...)

\method{Truncate}{matrix}(x, ncol = 10, nrow = 15, ...)

\method{Truncate}{data.frame}(x, ncol = 10, nrow = 15, ...)

\method{Truncate}{list}(x, ncol = 10, nrow = 15, ...)

\method{Truncate}{default}(x, ncol = 10, nrow = 15, ...)
}
\arguments{
\item{x}{An R object (usually a \code{data.frame}, \code{matrix}, or a
\code{list} that contains >= 1 \code{data.frame}/\code{matrix}.}

\item{ncol}{The number of columns (including 3 columns displayed as ellipsis)
to display.}

\item{nrow}{The number of rows (including 3 rows displayed as ellipsis) to
display.}

\item{\ldots}{Other arguments passed to \code{Truncate}.}
}
\value{
Prints a truncated display of the input.
}
\description{
A dispaly printing function that trunates \code{data.frame}/\code{matrix} by
adding ellipsis \ldots to take the place of non-displayed columns and rows.

\code{Truncate.matrix} - matrix method for \code{Truncate}

\code{Truncate.data.frame} - data.frame method for \code{Truncate}

\code{Truncate.list} - list method for \code{Truncate}

\code{Truncate.default} - Default method for \code{Truncate}
}
\note{
This function is intended for truncated display.  Please do not use to
gernarate your \href{http://sscce.org/}{SSCCE}
(Short, Self Contained, Correct (Compilable), Example).
}
\examples{
Truncate(mtcars)
Truncate(mtcars, nrow = 6)
Truncate(mtcars, ncol = 6)
Truncate(CO2)
Truncate(lm(mpg~hp+am, data=mtcars))
Truncate(1:10)
Truncate(mtcars[1:5, 1:5])
}
\keyword{truncate}

