\name{soaddnas}
\alias{soaddnas}
\title{Add random NAs to each column of a data frame}
\usage{
soaddnas(object, columns = "all", narate = 0.1)
}
\arguments{
  \item{object}{data.frame to which NAs should be added}

  \item{columns}{nu}
}
\description{
Add random NAs to each column of a data frame
}
\author{
Sebastian Campbell
}

